add embedding code

+++ Bi-LSTM/bi_lstm_model.py
@@ -13,6 +13,6 @@ class bi_lstm():
         lstm_bw_cell = tf.nn.rnn_cell.BasicLSTMCell(FLAGS.n_hidden)
         lstm_bw_cell = tf.nn.rnn_cell.DropoutWrapper(lstm_bw_cell, output_keep_prob=0.7)
 
-        outputs, _, _ = tf.nn.bidirectional_dynamic_rnn(lstm_fw_cell, lstm_bw_cell, input_data)
+        outputs, _, _ = tf.nn.bidirectional_dynamic_rnn(lstm_fw_cell, lstm_bw_cell, input_data, dtype=tf.int32)
 
         return tf.matmul(outputs[-1], weights['out'] + biases['out'])
